Removed Nav

+++ app/src/main/java/com/example/notes/NotesViewModel.kt
@@ -2,7 +2,6 @@ package com.example.notes
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
@@ -11,7 +10,7 @@ class NotesViewModel(private val repository: NotesRepository) : ViewModel(){
     val allNotes: LiveData<List<Notes>> = repository.allNotes.asLiveData()
     val favNotes: LiveData<List<Notes>> = repository.favNotes.asLiveData()
 
-    fun getNoteById(id: Int): LiveData<Notes> = repository.getNoteById(id).asLiveData()
+    fun getNoteById(id: Long): LiveData<Notes> = repository.getNoteById(id).asLiveData()
 
     fun insert(notes: Notes) = viewModelScope.launch {
         repository.insert(notes)
